+++ examples/api/multimodal/thread-openai/app.py
@@ -1,8 +1,9 @@
 from typing import Optional
 
 from agentkernel.api import RESTAPI
+from agentkernel.auth import Authoriser
 from agentkernel.openai import OpenAIModule
-from agentkernel.thread import AgentThreadRequestHandler, Authoriser
+from agentkernel.thread import AgentThreadRequestHandler
 from agents import Agent
 
 general_agent = Agent(
